fix Duration type mismatch bug in

+++ csharp/Adapter/Microsoft.Spark.CSharp/Proxy/Ipc/DStreamIpcProxy.cs
@@ -40,7 +40,7 @@ namespace Microsoft.Spark.CSharp.Proxy.Ipc
         public IDStreamProxy Window(int windowSeconds, int slideSeconds = 0)
         {
             string windowId = null;
-            var windowDurationReference = SparkCLRIpcProxy.JvmBridge.CallConstructor("org.apache.spark.streaming.Duration", new object[] { windowSeconds * 1000 });
+            var windowDurationReference = SparkCLRIpcProxy.JvmBridge.CallConstructor("org.apache.spark.streaming.Duration", new object[] { (long)windowSeconds * 1000 });
 
             if (slideSeconds <= 0)
             {
@@ -48,7 +48,7 @@ namespace Microsoft.Spark.CSharp.Proxy.Ipc
                 return new DStreamIpcProxy(new JvmObjectReference(windowId));
             }
 
-            var slideDurationReference = SparkCLRIpcProxy.JvmBridge.CallConstructor("org.apache.spark.streaming.Duration", new object[] { slideSeconds * 1000 });
+            var slideDurationReference = SparkCLRIpcProxy.JvmBridge.CallConstructor("org.apache.spark.streaming.Duration", new object[] { (long)slideSeconds * 1000 });
             windowId = (string)SparkCLRIpcProxy.JvmBridge.CallNonStaticJavaMethod(javaDStreamReference, "window", new object[] { windowDurationReference, slideDurationReference });
 
             return new DStreamIpcProxy(new JvmObjectReference(windowId));
